cleanup: move add todo to a different class
